+++ app/src/main/res/layout/fragment_notice_tab_event.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".NoticeTabEventFragment" >
+    tools:context=".notice.view.NoticeTabEventFragment" >
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recyclerview_notice_tab_event"
