+++ app/src/main/res/layout/fragment_notice_tab_notice.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".NoticeTabNoticeFragment">
+    tools:context=".notice.view.NoticeTabNoticeFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recyclerview_notice_tab_notice"
